moved the carousel under component and moved the portfolio and services pages under main content
