+++ frontend/src/components/carousel/Carousel.style.tsx
