+++ frontend/src/components/mainContent/portfolio/Portfolio.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MainContent } from '../mainContent/homeContent/HomeContent.style';
+import { MainContent } from '../homeContent/HomeContent.style';
 
 const Portfolio: React.FC = () => {
     return <MainContent>Portfolio</MainContent>;
